requirements.txt
Flask
Flask-Bcrypt
Flask-Login
python-dotenv
sqlite3
